relocate Script related assets
